requirements.txt
pandas
openpyxl
fastapi
uvicorn
gunicorn
python-dotenv
langgraph
langgraph-checkpoint-sqlite
langgraph-prebuilt
langchain-core
langchain-google-genai
streamlit
requests
langfuse
pydantic
sqlalchemy
psycopg2-binary
